refactor(recipe): update CartIconToggle import path
Move CartIconToggle from src/components to src/components/recipe
to better reflect its scope and organization in the project structure.
